requirements.txt
streamlit==1.18.1
PyPDF2==3.0.1
langchain==0.0.7
langchain_google_genai==0.0.1
google-generativeai==0.0.1
python-dotenv==1.0.0
tabula-py==2.5.0
fitz==1.0.0
ipython==8.10.0
Pillow==9.0.0
faiss-cpu==1.7.3
huggingface-hub==0.15.0
sentence-transformers==2.2.2
